requirements.txt
attrs
cached-property
future
itsdangerous
pyrsistent
Werkzeug
